computer_vision/ocr/requirements.txt
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
